fix kassa

+++ src/main/kotlin/com/gamestore/gamestorebackendkotlin/robokassa/dto/PaymentInput.kt
@@ -1,7 +1,7 @@
 package com.gamestore.gamestorebackendkotlin.robokassa.dto
 
 data class PaymentInput(
-    val OutSum: Double,
+    val OutSum: Int,
     val InvId: Int,
     val SignatureValue:String
 )
